Reintroduce the NoMaterializer for Play testkit purposes but in the stream testkit
