+++ akka-stream-tests/src/test/scala/akka/stream/impl/fusing/GraphInterpreterSpecKit.scala
@@ -4,10 +4,6 @@
 
 package akka.stream.impl.fusing
 
-import akka.actor.ActorSystem
-import akka.actor.Cancellable
-import akka.actor.Props
-import akka.annotation.InternalApi
 import akka.event.Logging
 import akka.stream.Supervision.Decider
 import akka.stream._
@@ -20,63 +16,10 @@ import akka.stream.impl.fusing.GraphInterpreter.{
 import akka.stream.stage.{ GraphStage, GraphStageLogic, InHandler, OutHandler, _ }
 import akka.stream.testkit.StreamSpec
 import akka.stream.testkit.Utils.TE
+import akka.stream.testkit.NoMaterializer
 import com.github.ghik.silencer.silent
 
 import scala.collection.{ Map => SMap }
-import scala.concurrent.ExecutionContextExecutor
-import scala.concurrent.duration.FiniteDuration
-
-/**
- * INTERNAL API
- */
-@InternalApi
-private[akka] object NoMaterializer extends Materializer {
-  override def withNamePrefix(name: String): Materializer =
-    throw new UnsupportedOperationException("NoMaterializer cannot be named")
-  override def materialize[Mat](runnable: Graph[ClosedShape, Mat]): Mat =
-    throw new UnsupportedOperationException("NoMaterializer cannot materialize")
-  override def materialize[Mat](runnable: Graph[ClosedShape, Mat], defaultAttributes: Attributes): Mat =
-    throw new UnsupportedOperationException("NoMaterializer cannot materialize")
-
-  override def executionContext: ExecutionContextExecutor =
-    throw new UnsupportedOperationException("NoMaterializer does not provide an ExecutionContext")
-
-  def scheduleOnce(delay: FiniteDuration, task: Runnable): Cancellable =
-    throw new UnsupportedOperationException("NoMaterializer cannot schedule a single event")
-
-  def schedulePeriodically(initialDelay: FiniteDuration, interval: FiniteDuration, task: Runnable): Cancellable =
-    throw new UnsupportedOperationException("NoMaterializer cannot schedule a repeated event")
-
-  override def scheduleWithFixedDelay(
-      initialDelay: FiniteDuration,
-      delay: FiniteDuration,
-      task: Runnable): Cancellable =
-    throw new UnsupportedOperationException("NoMaterializer cannot scheduleWithFixedDelay")
-
-  override def scheduleAtFixedRate(
-      initialDelay: FiniteDuration,
-      interval: FiniteDuration,
-      task: Runnable): Cancellable =
-    throw new UnsupportedOperationException("NoMaterializer cannot scheduleAtFixedRate")
-
-  override def shutdown(): Unit = throw new UnsupportedOperationException("NoMaterializer cannot shutdown")
-
-  override def isShutdown: Boolean = throw new UnsupportedOperationException("NoMaterializer cannot shutdown")
-
-  override def system: ActorSystem =
-    throw new UnsupportedOperationException("NoMaterializer does not have an actorsystem")
-
-  override private[akka] def logger = throw new UnsupportedOperationException("NoMaterializer does not have a logger")
-
-  override private[akka] def supervisor =
-    throw new UnsupportedOperationException("NoMaterializer does not have a supervisor")
-
-  override private[akka] def actorOf(context: MaterializationContext, props: Props) =
-    throw new UnsupportedOperationException("NoMaterializer cannot spawn actors")
-
-  override def settings: ActorMaterializerSettings =
-    throw new UnsupportedOperationException("NoMaterializer does not have settings")
-}
 
 @silent
 object GraphInterpreterSpecKit {
